ball functions like a pong ball, all that's really left is adjusting numbers to make it feel as good as possible. In terms of functionality, the ball is done.

+++ game/Ball.hpp
@@ -11,10 +11,10 @@ using namespace std;
 
 class Ball {
 private:
-    int xCoord = 0;
-    int yCoord = 0;
-    int speedX = 0;
-    int speedY = 0;
+    double xCoord = 0;
+    double yCoord = 0;
+    double speedX = 0;
+    double speedY = 0;
 public:
     // Default Constructor
     Ball() = default;
@@ -22,16 +22,16 @@ public:
     Ball(int xCoord, int yCoord):xCoord(xCoord),yCoord(yCoord){};
 
     // Getters
-    int getXCoord(){ return xCoord; }
-    int getYCoord(){ return yCoord; }
-    int getXSpeed(){return speedX;}
-    int getYSpeed(){return speedY;}
+    double getXCoord(){ return xCoord; }
+    double getYCoord(){ return yCoord; }
+    double getXSpeed(){return speedX;}
+    double getYSpeed(){return speedY;}
 
     // Setters
-    void setXCoord(int X){ xCoord = X; }
-    void setYCoord(int Y){ yCoord = Y; }
-    void setXSpeed(int X){ speedX = X; }
-    void setYSpeed(int Y){ speedY = Y; }
+    void setXCoord(double X){ xCoord = X; }
+    void setYCoord(double Y){ yCoord = Y; }
+    void setXSpeed(double X){ speedX = X; }
+    void setYSpeed(double Y){ speedY = Y; }
 
 };
